requirements.txt
Flask==1.1.2
RPi.GPIO==0.7.0; 'arm' in platform_machine
fake-rpi==0.7.1; 'x86' in platform_machine
